test test test

+++ api/quotes/delete.php
@@ -13,19 +13,25 @@
     $data = json_decode(file_get_contents("php://input"));
 
     // Set ID to update
-    $quote->id = $data->id;
+    $quote->id = isset($data->id) ? $data->id : null;
 
     //Delete quote
-    if($quote->delete()){
-        //Create array
-        $quote_arr = array(
-            'id'            => $quote->id,
-        );
+    if(isset($quote->id)){
+        if($quote->delete()){
+            //Create array
+            $quote_arr = array(
+                'id'            => $quote->id,
+            );
 
-        // Make JSON
-        print_r(json_encode($quote_arr));
+            // Make JSON
+            print_r(json_encode($quote_arr));
+        }else{
+            echo json_encode(
+                array('message' => 'No Quotes Found')
+            );  
+        }
     }else{
         echo json_encode(
             array('message' => 'No Quotes Found')
-        );  
+        ); 
     }
